Avoid downloading existing files

+++ src/AI_WQ_package/retrieve_training_data.py
@@ -32,23 +32,23 @@ def retrieve_annual_training_data(year,variable,password,local_destination=None)
         else:
             local_filename = f'{local_destination}/{variable}_sevenday_WEEKLYSUM_{year}.nc'
 
-    # log onto FTP session
-    session = ftplib.FTP('ftp.ecmwf.int','ai_weather_quest',password)
-    if local_destination == None:
-        remote_path = f'/training_data/{local_filename}'
+    if not os.path.exists(local_filename):
+        # log onto FTP session
+        session = ftplib.FTP('ftp.ecmwf.int','ai_weather_quest',password)
+        if local_destination == None:
+            remote_path = f'/training_data/{local_filename}'
+        else:
+            if variable == 'tas' or variable == 'mslp':
+                remote_path = f'/training_data/{variable}_sevenday_WEEKLYMEAN_{year}.nc'
+            elif variable == 'pr':
+                remote_path = f'/training_data/{variable}_sevenday_WEEKLYSUM_{year}.nc'
+        with open(local_filename, 'wb') as f:
+            session.retrbinary(f"RETR {remote_path}", f.write)
+        print(f"File '{remote_path}' has been downloaded successfully.")
+        session.quit()
     else:
-        if variable == 'tas' or variable == 'mslp':
-            remote_path = f'/training_data/{variable}_sevenday_WEEKLYMEAN_{year}.nc'
-        elif variable == 'pr':
-            remote_path = f'/training_data/{variable}_sevenday_WEEKLYSUM_{year}.nc'
+        print(f"File '{local_filename}' already exists. Skipping download.")
     
-    # retrieve the full year file 
-    with open(local_filename,'wb') as f:
-        session.retrbinary(f"RETR {remote_path}", f.write)
-
-    print(f"File '{remote_path}' has been downloaded to successfully to '{local_filename}'.")
-
-    session.quit()
     # open file using xarray. # removes time bounds
     full_year_obs = xr.open_dataset(local_filename).squeeze()
     return full_year_obs
